legacy/python/requirements.txt
mcp
ollama
asyncio
requests
python-dotenv
